% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bools_to_int.R
\name{bools_to_int}
\alias{bools_to_int}
\title{Convert a vector of booleans to an integer}
\usage{
bools_to_int(bools)
}
\arguments{
\item{bools}{a vector of booleans,
with at least 1 element}
}
\value{
a integer
}
\description{
Convert a vector of booleans to an integer,
using a big-endian conversion, i.e.
the last bool encodes for a 1,
the one-but-last bool encodes for a 2,
etcetera.
}
\examples{
bools_to_int(FALSE) # 0
bools_to_int(TRUE) # 1
bools_to_int(c(TRUE, FALSE)) # 2
bools_to_int(c(TRUE, TRUE)) # 3
bools_to_int(c(TRUE, FALSE, FALSE)) # 4
bools_to_int(c(TRUE, FALSE, TRUE)) # 5
bools_to_int(c(TRUE, TRUE, FALSE)) # 6
bools_to_int(c(TRUE, TRUE, TRUE)) # 7
}
\author{
Richèl J.C. Bilderbeek
}
