% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_group_id_to_writer_names.R
\name{convert_group_id_to_writer_names}
\alias{convert_group_id_to_writer_names}
\title{Convert a group ID to their writers' names}
\usage{
convert_group_id_to_writer_names(group_id)
}
\arguments{
\item{group_id}{a group ID. Group IDs may differ between versions
of this package}
}
\value{
a vector of writers
}
\description{
Convert a group ID to their writers' names
}
\examples{
convert_group_id_to_writer_names(0) # empty
convert_group_id_to_writer_names(1) # "Tracy Atkins"
convert_group_id_to_writer_names(2) # "Ronald Buersens"
convert_group_id_to_writer_names(3) # c("Ronald Buersens", "Tracy Atkins")
}
\author{
Richèl J.C. Bilderbeek
}
