% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cds.R
\name{get_cds}
\alias{get_cds}
\title{Get the K3 CDs}
\usage{
get_cds()
}
\value{
a tibble with the K3 CDs:
\itemize{
\item \code{cd_title}: title of the CD
\item \code{formation}: the formation number,
where '1' denotes the first formation
(see \link{get_formations} to get information on the formations)
\item \code{year}: the release year of the CD
}
}
\description{
Get the K3 CDs
}
\examples{
get_cds()
}
\author{
Richèl J.C. Bilderbeek
}
