% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_composer_names.R
\name{get_composer_names}
\alias{get_composer_names}
\title{Get the name of all composers}
\usage{
get_composer_names()
}
\value{
an vector of all composer names,
ordered alphabetically
}
\description{
Get the name of all composers
}
\examples{
get_composer_names()
}
\author{
Richèl J.C. Bilderbeek
}
