% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_composers.R
\name{get_composers}
\alias{get_composers}
\title{Get the K3 songs' composers}
\usage{
get_composers()
}
\value{
a tibble with the K3 songs' composers. Column names are:
\itemize{
\item \code{song_title}: title of the song
\item \code{composer}: name of a composer
}

For songs with multiple composers, multiple rows are used.
}
\description{
Get the K3 songs' composers
}
\examples{
get_composers()
}
\author{
Richèl J.C. Bilderbeek
}
