% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formations.R
\name{get_formations}
\alias{get_formations}
\title{Get the K3 formations}
\usage{
get_formations()
}
\value{
a tibble with the K3 formations. Column names are:
\itemize{
\item \code{formation}: the formation number, where 1 is the first formation
\item \code{member_1}: the first member of that formation
\item \code{member_2}: the second member of that formation
\item \code{member_3}: the third member of that formation
}

The member names are sorted alphabetically
}
\description{
Get the K3 formations
}
\examples{
get_formations()
}
\author{
Richèl J.C. Bilderbeek
}
