% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heyahmama_path.R
\name{get_heyahmama_path}
\alias{get_heyahmama_path}
\title{Get the path to a file bundled with this package}
\usage{
get_heyahmama_path(filename)
}
\value{
the full path to a file bundled with this package
}
\description{
Get the path to a file bundled with this package
}
\examples{
get_heyahmama_path("formations.csv")
}
\author{
Richèl J.C. Bilderbeek
}
