% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_released_songs.R
\name{get_released_songs}
\alias{get_released_songs}
\title{Get the K3 released songs}
\usage{
get_released_songs()
}
\value{
a tibble with released K3 songs.

Column names are:
\itemize{
\item \code{cd_title}: title of the CD
\item \code{cd_number}: the CD number. For single-CD releases, this will be 1.
\item \code{track_number}: track number on that CD, e.g.
1 for the first track
\item \code{song_title}: title of the song
\item \code{duration}: duration of this recording of the song,
as indicated on the CD.
Some songs, e.g. 'Heyah mama', have different durations for
different recordings
}
}
\description{
Get the K3 released songs
}
\examples{
get_released_songs()
}
\author{
Richèl J.C. Bilderbeek
}
