% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_filename.R
\name{get_song_filename}
\alias{get_song_filename}
\title{Get a K3 songs' filename}
\usage{
get_song_filename(song_title)
}
\arguments{
\item{song_title}{title of a K3 song.
Capitalization of this value is ignored.}
}
\value{
a filename
}
\description{
Get a K3 songs' filename
}
\examples{
get_song_filename("Heyah Mama")
}
\author{
Richèl J.C. Bilderbeek
}
