% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_lyrics.R
\name{get_song_lyrics}
\alias{get_song_lyrics}
\title{Get a K3 songs' lyrics}
\usage{
get_song_lyrics(song_title)
}
\arguments{
\item{song_title}{title of a K3 song.
Capitalization of this value is ignored.}
}
\value{
a text. Will \link{stop} if the song
cannot be found
}
\description{
Get a K3 songs' lyrics
}
\examples{
get_song_lyrics("Heyah Mama")
}
\author{
Richèl J.C. Bilderbeek
}
