% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_titles.R
\name{get_song_titles}
\alias{get_song_titles}
\title{Get the titles of all K3 songs, sorted alphabetically}
\usage{
get_song_titles()
}
\value{
a vector with titles of all K3 songs
}
\description{
Get the titles of all K3 songs, sorted alphabetically
}
\examples{
get_song_titles()
}
\author{
Richèl J.C. Bilderbeek
}
