% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_songs.R
\name{get_songs}
\alias{get_songs}
\title{Get the K3 songs}
\usage{
get_songs()
}
\value{
a tibble with the K3 songs
}
\description{
Get the K3 songs
}
\examples{
get_songs()
}
\author{
Richèl J.C. Bilderbeek
}
