% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_writer_names.R
\name{get_writer_names}
\alias{get_writer_names}
\title{Get the name of all writers}
\usage{
get_writer_names()
}
\value{
an vector of all writer names,
ordered alphabetically
}
\description{
Get the name of all writers
}
\examples{
get_writer_names()
}
\author{
Richèl J.C. Bilderbeek
}
