% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_writers.R
\name{get_writers}
\alias{get_writers}
\title{Get the K3 songs' writers}
\usage{
get_writers()
}
\value{
a tibble with the K3 songs' writers. Column names are:
\itemize{
\item \code{song_title}: title of the song
\item \code{writer}: name of a writer
}

For songs with multiple writers, multiple rows are used.
}
\description{
Get the K3 songs' writers
}
\examples{
get_writers()
}
\author{
Richèl J.C. Bilderbeek
}
