% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_writers_groups.R
\name{get_writers_groups}
\alias{get_writers_groups}
\title{Get the different writers groups}
\usage{
get_writers_groups()
}
\value{
an list with two elements:
\itemize{
\item \code{song_title_to_group_id}: a tibble with columns
\code{song_title} and \code{group_id}.
A \code{group_id} of zero denotes that the writers are
unknown
\item \code{group_id_to_writer_name}: a tibble with columns
\code{group_id} and \code{writer_name}.
A \code{group_id} of zero denotes that the writers are
unknown
}
}
\description{
Get the different writers groups
}
\examples{
get_writers_groups()
}
\author{
Richèl J.C. Bilderbeek
}
