% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_to_bools.R
\name{int_to_bools}
\alias{int_to_bools}
\title{Convert an integer to a a vector of booleans}
\usage{
int_to_bools(i)
}
\arguments{
\item{i}{an integer}
}
\value{
a vector of booleans
}
\description{
Convert an integer to a a vector of booleans
using a big-endian conversion, i.e.
the last bool encodes for a 1,
the one-but-last bool encodes for a 2,
etcetera.
}
\examples{
int_to_bools(0) # FALSE
int_to_bools(1) # TRUE
int_to_bools(2) # c(TRUE, FALSE)
int_to_bools(3) # c(TRUE, TRUE)
int_to_bools(4) # c(TRUE, FALSE, FALSE)
}
\author{
Richèl J.C. Bilderbeek
}
