% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_writer_names_to_group_id.R
\name{convert_writer_names_to_group_id}
\alias{convert_writer_names_to_group_id}
\title{Convert writers names to a group ID}
\usage{
convert_writer_names_to_group_id(writer_names)
}
\arguments{
\item{writer_names}{the names of the writers}
}
\value{
a group ID
}
\description{
Convert writers names to a group ID
}
\examples{
convert_writer_names_to_group_id("Alain Vande Putte") # 1024
convert_writer_names_to_group_id("Tracy Atkins") # 1
convert_writer_names_to_group_id(c("Alain Vande Putte", "Tracy Atkins")) # 1025
}
\author{
Richèl J.C. Bilderbeek
}
