% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formations.R
\name{get_formations}
\alias{get_formations}
\title{Get the K3 formations}
\usage{
get_formations()
}
\value{
a tibble with the K3 formations
}
\description{
Get the K3 formations
}
\examples{
get_formations()
}
\author{
Richèl J.C. Bilderbeek
}
