% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_songs.R
\name{get_songs}
\alias{get_songs}
\title{Get the K3 songs}
\usage{
get_songs()
}
\value{
a tibble with the K3 songs. Column names are:
\itemize{
\item \code{cd_title}: title of the CD
\item \code{cd_number}: the CD number. For single-CD releases, this will be 1.
\item \code{track_number}: track number on that CD, e.g.
1 for the first track
\item \code{song_title}: title of the song
\item \code{duration}: duration of this recording of the song.
For some songs, e.g. 'Heyah mama', recordings differ
}
}
\description{
Get the K3 songs
}
\examples{
get_songs()
}
\author{
Richèl J.C. Bilderbeek
}
