% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_songs.R
\name{get_songs}
\alias{get_songs}
\title{Get the K3 songs}
\usage{
get_songs()
}
\value{
a tibble with the K3 songs. Column names are:
\itemize{
\item \code{cd_title}: title of the CD
\item \code{track_number}: track number on that CD, e.g.
1 for the first track
\item \code{song_title}: title of the song
}
}
\description{
Get the K3 songs
}
\examples{
get_songs()
}
\author{
Richèl J.C. Bilderbeek
}
